+++ .travis/deploy.sh
@@ -3,6 +3,6 @@
 brew update > /dev/null
 brew outdated carthage || brew upgrade carthage
 make build
-yarn install
+yarn
 yarn build
 yarn pack