+++ rollup.config.js
@@ -1,7 +1,6 @@
 import commonjs from 'rollup-plugin-commonjs';
 import uglify from 'rollup-plugin-uglify';
 import resolve from 'rollup-plugin-node-resolve';
-import shim from 'rollup-plugin-shim';
 import typescript from 'rollup-plugin-typescript2';
 
 const defaultPlugins = [
@@ -15,9 +14,6 @@ const defaultPlugins = [
     browser: true,
   }),
   commonjs(),
-  shim({
-    domain: `export var active = false;`,
-  }),
 ];
 
 const defaultMinPlugins = [
@@ -31,9 +27,6 @@ const defaultMinPlugins = [
     browser: true,
   }),
   commonjs(),
-  shim({
-    domain: `export var active = false;`,
-  }),
   uglify(),
 ];