+++ src/js/sdk.ts
@@ -1,33 +1,31 @@
+import { Integrations as BrowserIntegrations } from '@sentry/browser';
 import { initAndBind } from '@sentry/core';
 import { CordovaOptions } from './backend';
-import { CordovaFrontend } from './frontend';
-import { setExtraContext as shimSetExtraContext } from '@sentry/shim';
-
-export {
-  addBreadcrumb,
-  captureEvent,
-  captureException,
-  captureMessage,
-  clearScope,
-  popScope,
-  pushScope,
-  setExtraContext,
-  setTagsContext,
-  setUserContext,
-} from '@sentry/shim';
+import { CordovaClient } from './client';
+import { Cordova, Release } from './integrations';
+import { Scope } from '@sentry/hub';
+import { configureScope } from '@sentry/minimal';
 
 export function init(options: CordovaOptions): void {
-  initAndBind(CordovaFrontend, options);
+  initAndBind(CordovaClient, options, [
+    new BrowserIntegrations.OnError(),
+    new BrowserIntegrations.OnUnhandledRejection(),
+    new BrowserIntegrations.FunctionToString(),
+    new BrowserIntegrations.TryCatch(),
+    new BrowserIntegrations.Breadcrumbs(),
+    new Cordova(),
+    new Release(),
+  ]);
 }
 
 export function setRelease(release: string): void {
-  shimSetExtraContext({ __sentry_release: release });
+  configureScope((scope: Scope) => {
+    scope.setExtra('__sentry_release', release);
+  });
 }
 
 export function setDist(dist: string): void {
-  shimSetExtraContext({ __sentry_dist: dist });
-}
-
-export function setVersion(version: string): void {
-  shimSetExtraContext({ __sentry_version: version });
+  configureScope((scope: Scope) => {
+    scope.setExtra('__sentry_dist', dist);
+  });
 }