+++ src/js/sentry-cordova.ts
@@ -1,16 +1,30 @@
-export { CordovaFrontend } from './frontend';
+export {
+  Breadcrumb,
+  Integration,
+  Request,
+  SdkInfo,
+  SentryEvent,
+  SentryException,
+  Severity,
+  StackFrame,
+  Stacktrace,
+  Thread,
+  User,
+} from '@sentry/types';
+
 export {
   addBreadcrumb,
-  captureEvent,
-  captureException,
   captureMessage,
-  init,
-  popScope,
-  pushScope,
-  setDist,
-  setUserContext,
-  setExtraContext,
-  setRelease,
-  setTagsContext,
-  setVersion,
-} from './sdk';
+  captureException,
+  captureEvent,
+  configureScope,
+} from '@sentry/minimal';
+
+export { getDefaultHub, getHubFromCarrier, Hub, Scope } from '@sentry/hub';
+
+export { CordovaBackend, CordovaOptions } from './backend';
+export { CordovaClient } from './client';
+export { init, setDist, setRelease } from './sdk';
+
+import * as Integrations from './integrations';
+export { Integrations };